refactor(update): extract Playback dispatch to playback.rs (commit 2/N)
Pull the 120-line Message::Playback(...) inline match out of update/mod.rs
into a new dispatch_playback method on the existing playback.rs file.
The Tick arm's pre-handler housekeeping (stale-loading watchdog +
active-progress toast polling, ~80 lines) moves with it inline since
that's strictly per-tick state-poll work with no other natural home.

mod.rs match arm:
    Message::Playback(msg) => self.dispatch_playback(msg),

cargo build / clippy --all-targets -D warnings / nightly fmt --check /
test --workspace all clean (662 + 368 + 1 doctest).
